CHC solver: move solver_types.h methods into solver_types.cpp
This moves the body of the methods defined in the header file into a
corresponding .cpp file.
